Codigos de Respuesta Http implementado en los endpoint

+++ src/main/java/my_api/controllers/userController.java
@@ -1,6 +1,8 @@
 package my_api.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -10,10 +12,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
-import java.util.Optional;
-import java.util.Date;
 
+import java.util.Date;
+import java.util.Optional;
 import my_api.models.User;
 
 import my_api.repositories.userRepository;
@@ -28,46 +29,67 @@ public class userController {
 
     // Endpoint para obtener los usuarios
     @GetMapping
-    public List<User> getUsers()
+    public ResponseEntity<?> getUsers()
     {
-        return userRepository.findAll();
+        if (userRepository.findAll().isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("No hay usuarios encontrado");
+        }
+        return ResponseEntity.ok(userRepository.findAll());
     }
 
     // Endpoint para crear usuario
     @PostMapping
-    public User postUser(@RequestBody User user)
+    public ResponseEntity<User> postUser(@RequestBody User user)
     {
         user.setCreatedAt(new Date());
-        return userRepository.save(user);
+        return ResponseEntity.status(HttpStatus.CREATED)
+        .body(userRepository.save(user));
     }
 
     // Endpoint para buscar usuarios por id
 
     @GetMapping("/{id}")
-    public Optional<User> getUser(@PathVariable Long id)
+    public ResponseEntity<?> getUser(@PathVariable Long id)
     {
-        return userRepository.findById(id);
+        if (userRepository.findById(id).isEmpty())
+        {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+            .body("No se encontro el usuario con el id: " + id);
+        }
+        return ResponseEntity.ok(userRepository.findById(id));
     }
 
     // Endpoint para actualizar usuario
     @PutMapping("/{id}")
-    public User putUser(@PathVariable Long id, @RequestBody User userBody)
+    public ResponseEntity<?> putUser(@PathVariable Long id, @RequestBody User userBody)
     {
-        User user = userRepository.findById(id).orElseThrow(() -> new RuntimeException("User not found"));
+        Optional<User> OptionalUser = userRepository.findById(id);
+
+        if (OptionalUser.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+            .body("No se encontro el usuario con el id: " + id);
+        }
+        User user = OptionalUser.get();
         user.setUsername(userBody.getUsername());
         user.setEmail(userBody.getEmail());
         user.setPassword(userBody.getPassword());
         user.setFirstname(userBody.getFirstname());
         user.setLastname(userBody.getLastname());
         user.setRoles(userBody.getRoles());
-        return userRepository.save(user);
+        userRepository.save(user);
+        return ResponseEntity.noContent().build();
     }
 
     // Endpoint para eliminar usuario por id
     @DeleteMapping("/{id}")
-    public void deleteUser(@PathVariable Long id)
+    public ResponseEntity<?> deleteUser(@PathVariable Long id)
     {
+        if (userRepository.findById(id).isEmpty()) 
+        {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Usuario no existe");    
+        }
         userRepository.deleteById(id);
+        return ResponseEntity.noContent().build();
     }
